pyproject.toml
[tool.poetry]
name = "ptax"
version = "0.1.0"
description = ""
authors = ["Thiago Oliveira <thiceconelo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
lxml = "^4.9.0"
html-table-parser-python3 = "^0.2.0"
parsel = "^1.6.0"
tweepy = "^4.10.0"
python-crontab = "^2.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
